# docs/openapi.json
{
  "components": {
    "schemas": {
      "AddExerciseRequest": {
        "properties": {
          "exercise_id": {
            "minimum": 1.0,
            "title": "Exercise Id",
            "type": "integer"
          }
        },
        "required": [
          "exercise_id"
        ],
        "title": "AddExerciseRequest",
        "type": "object"
      },
      "AddSetRequest": {
        "properties": {
          "reps": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Reps",
            "type": "integer"
          },
          "weight": {
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "weight",
          "reps"
        ],
        "title": "AddSetRequest",
        "type": "object"
      },
...
      "DeleteWorkoutResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "workout_id": {
            "title": "Workout Id",
            "type": "integer"
          }
        },
        "required": [
          "deleted",
          "workout_id"
        ],
        "title": "DeleteWorkoutResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoadMetricsResponse": {
        "properties": {
          "back_stress_score": {
            "title": "Back Stress Score",
            "type": "number"
          },
          "compound_score": {
            "title": "Compound Score",
            "type": "number"
          },
          "intensity_score": {
            "title": "Intensity Score",
            "type": "number"
          },
          "load_label": {
            "title": "Load Label",
            "type": "string"
          },
          "load_score": {
            "title": "Load Score",
            "type": "number"
          }
        },
        "required": [
          "load_score",
          "load_label",
          "compound_score",
          "back_stress_score"
        ],
        "title": "LoadMetricsResponse",
        "type": "object"
      },
...
      "UpdateSetRequest": {
        "properties": {
          "reps": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Reps",
            "type": "integer"
          },
          "weight": {
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "title": "UpdateSetRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkoutAnalysisExerciseResponse": {
        "properties": {
          "best_e1rm": {
            "title": "Best E1Rm",
            "type": "number"
          },
          "best_e1rm_set": {
            "title": "Best E1Rm Set",
            "type": "object"
          },
          "best_set": {
            "title": "Best Set",
            "type": "object"
          },
          "exercise_id": {
            "title": "Exercise Id",
            "type": "integer"
          },
          "exercise_name": {
            "title": "Exercise Name",
            "type": "string"
          },
          "pr_flags": {
            "items": {
              "type": "string"
            },
            "title": "Pr Flags",
            "type": "array"
          }
        },
        "required": [
          "exercise_id",
          "exercise_name",
          "pr_flags"
        ],
        "title": "WorkoutAnalysisExerciseResponse",
        "type": "object"
      },
      "WorkoutAnalysisPrResponse": {
        "properties": {
          "exercise_name": {
            "title": "Exercise Name",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "exercise_name",
          "type"
        ],
        "title": "WorkoutAnalysisPrResponse",
        "type": "object"
      },
      "WorkoutAnalysisResponse": {
        "properties": {
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/WorkoutAnalysisExerciseResponse"
            },
            "title": "Exercises",
            "type": "array"
          },
          "prs": {
            "items": {
              "$ref": "#/components/schemas/WorkoutAnalysisPrResponse"
            },
            "title": "Prs",
            "type": "array"
          }
        },
        "required": [
          "exercises",
          "prs"
        ],
        "title": "WorkoutAnalysisResponse",
        "type": "object"
      },
      "WorkoutCoreResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "integer"
          },
          "finished_at": {
            "title": "Finished At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "lower_back_pain": {
            "title": "Lower Back Pain",
            "type": "integer"
          },
          "session_rpe": {
            "title": "Session Rpe",
            "type": "integer"
          },
          "workout_date": {
            "title": "Workout Date",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workout_date",
          "created_at"
        ],
        "title": "WorkoutCoreResponse",
        "type": "object"
      },
      "WorkoutDetailResponse": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/WorkoutAnalysisResponse"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/WorkoutExerciseResponse"
            },
            "title": "Exercises",
            "type": "array"
          },
          "load_metrics": {
            "$ref": "#/components/schemas/LoadMetricsResponse"
          },
          "total_reps": {
            "title": "Total Reps",
            "type": "integer"
          },
          "total_sets": {
            "title": "Total Sets",
            "type": "integer"
          },
          "total_volume": {
            "title": "Total Volume",
            "type": "number"
          },
          "workout": {
            "$ref": "#/components/schemas/WorkoutCoreResponse"
          }
        },
        "required": [
          "workout",
          "exercises",
          "total_volume",
          "total_reps",
          "total_sets",
          "load_metrics",
          "analysis"
        ],
        "title": "WorkoutDetailResponse",
        "type": "object"
      },
      "WorkoutExerciseResponse": {
        "properties": {
          "configured_weights": {
            "items": {
              "type": "number"
            },
            "title": "Configured Weights",
            "type": "array"
          },
          "default_reps": {
            "title": "Default Reps",
            "type": "integer"
          },
          "default_weight": {
            "title": "Default Weight",
            "type": "number"
          },
          "exercise_id": {
            "title": "Exercise Id",
            "type": "integer"
          },
          "exercise_name": {
            "title": "Exercise Name",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "profile_key": {
            "title": "Profile Key",
            "type": "string"
          },
          "sets": {
            "items": {
              "$ref": "#/components/schemas/SetEntryResponse"
            },
            "title": "Sets",
            "type": "array"
          },
          "total_reps": {
            "title": "Total Reps",
            "type": "integer"
          },
          "total_sets": {
            "title": "Total Sets",
            "type": "integer"
          },
          "total_volume": {
            "title": "Total Volume",
            "type": "number"
          },
          "workout_exercise_id": {
            "title": "Workout Exercise Id",
            "type": "integer"
          }
        },
        "required": [
          "workout_exercise_id",
          "exercise_id",
          "exercise_name",
          "profile_key",
          "position",
          "sets",
          "total_sets",
          "total_reps",
          "total_volume",
          "default_weight",
          "default_reps",
          "configured_weights"
        ],
        "title": "WorkoutExerciseResponse",
        "type": "object"
      },
...
      "WorkoutUpdateRequest": {
        "properties": {
          "created_at": {
            "maxLength": 19,
            "minLength": 16,
            "title": "Created At",
            "type": "string"
          },
          "lower_back_pain": {
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Lower Back Pain",
            "type": "integer"
          },
          "session_rpe": {
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Session Rpe",
            "type": "integer"
          }
        },
        "title": "WorkoutUpdateRequest",
        "type": "object"
      },
      "WorkoutsResponse": {
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "workouts": {
            "items": {
              "$ref": "#/components/schemas/WorkoutSummaryResponse"
            },
            "title": "Workouts",
            "type": "array"
          }
        },
        "required": [
          "limit",
          "workouts"
        ],
        "title": "WorkoutsResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Training Log",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sets/{set_id}": {
      "delete": {
        "operationId": "delete_set_endpoint_api_v1_sets__set_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "title": "Set Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Set Endpoint",
        "tags": [
          "workouts"
        ]
      },
      "patch": {
        "operationId": "update_set_endpoint_api_v1_sets__set_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "title": "Set Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Set Endpoint",
        "tags": [
          "workouts"
        ]
      }
    },
...
    "/api/v1/workout-exercises/{workout_exercise_id}/sets": {
      "post": {
        "operationId": "add_workout_exercise_set_endpoint_api_v1_workout_exercises__workout_exercise_id__sets_post",
        "parameters": [
          {
            "in": "path",
            "name": "workout_exercise_id",
            "required": true,
            "schema": {
              "title": "Workout Exercise Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Workout Exercise Set Endpoint",
        "tags": [
          "workouts"
        ]
      }
    },
    "/api/v1/workout-exercises/{workout_exercise_id}/sets/duplicate": {
      "post": {
        "operationId": "duplicate_workout_exercise_set_endpoint_api_v1_workout_exercises__workout_exercise_id__sets_duplicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "workout_exercise_id",
            "required": true,
            "schema": {
              "title": "Workout Exercise Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Duplicate Workout Exercise Set Endpoint",
        "tags": [
          "workouts"
        ]
      }
    },
    "/api/v1/workouts": {
      "get": {
        "operationId": "get_workouts_api_v1_workouts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 500.0,
              "minimum": 1.0,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workouts",
        "tags": [
          "workouts"
        ]
      }
    },
    "/api/v1/workouts/{workout_id}": {
      "delete": {
        "operationId": "delete_workout_endpoint_api_v1_workouts__workout_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "title": "Workout Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Workout Endpoint",
        "tags": [
          "workouts"
        ]
      },
      "get": {
        "operationId": "get_workout_detail_api_v1_workouts__workout_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "title": "Workout Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workout Detail",
        "tags": [
          "workouts"
        ]
      },
      "patch": {
        "operationId": "update_workout_endpoint_api_v1_workouts__workout_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "title": "Workout Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Workout Endpoint",
        "tags": [
          "workouts"
        ]
      }
    },
    "/api/v1/workouts/{workout_id}/exercises": {
      "post": {
        "operationId": "add_workout_exercise_endpoint_api_v1_workouts__workout_id__exercises_post",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "title": "Workout Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddExerciseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Workout Exercise Endpoint",
        "tags": [
          "workouts"
        ]
      }
    },
    "/api/v1/workouts/{workout_id}/exercises/{workout_exercise_id}": {
      "delete": {
        "operationId": "delete_workout_exercise_endpoint_api_v1_workouts__workout_id__exercises__workout_exercise_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "title": "Workout Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "workout_exercise_id",
            "required": true,
            "schema": {
              "title": "Workout Exercise Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Workout Exercise Endpoint",
        "tags": [
          "workouts"
        ]
      }
    }
  }
}